+++ system/platform-core/src/test/java/org/platformlambda/core/models/SimplePoJo.java
@@ -20,7 +20,7 @@ package org.platformlambda.core.models;
 
 public class SimplePoJo {
 
-    public String name;
+    public String fullName;
     public String address;
     public int telephone;
 }